% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{estimated_FDR}
\alias{estimated_FDR}
\title{Function for estimating the FDR}
\usage{
estimated_FDR(t, p_vec, hat_pi_0, hat_F_0_t)
}
\arguments{
\item{t}{vector of thresholds for p-values}

\item{p_vec}{vector of 1D p-values}

\item{hat_pi_0}{scalar, \eqn{\hat{\pi}_0}}

\item{hat_F_0_t}{the null distribution, same size as t}
}
\value{
the estimated FDR, same size as t
}
\description{
Function for estimating the FDR
}
