% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{estimated_pi_0}
\alias{estimated_pi_0}
\title{Function for estimating the null proportion}
\usage{
estimated_pi_0(lambda, p_vec, hat_F_0_lambda)
}
\arguments{
\item{lambda}{scalar, auxiliary parameter}

\item{p_vec}{a vector of 1D p-values}

\item{hat_F_0_lambda}{\eqn{\hat{F}_0(\lambda)}}
}
\value{
\eqn{\hat\pi_0(\lambda)}
}
\description{
Function for estimating the null proportion
}
