% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nonparametric_estimated_F_0}
\alias{nonparametric_estimated_F_0}
\title{Nonparametric method for estimating the null distribution}
\usage{
nonparametric_estimated_F_0(t, p_vec)
}
\arguments{
\item{t}{vector of thresholds for p-values}

\item{p_vec}{vector of 1D p-values}
}
\value{
the null distribution, same size as t
}
\description{
Nonparametric method for estimating the null distribution
}
