% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM}
\alias{SIM}
\title{The main function in SIM paper}
\usage{
SIM(p_1, p_2, alpha, SIM_F_0_method = 1, method_lambda = 1, method_t = 1)
}
\arguments{
\item{p_1}{the primary p-values}

\item{p_2}{the auxiliary p-values}

\item{alpha}{the significance level}

\item{SIM_F_0_method}{method for estimating the null distrbution (1: parametric; 2: nonparametic method)}

\item{method_lambda}{method for searching the lambda used in null proportion (1: specified; 2: adaptive)}

\item{method_t}{method for searching the p-value threshold (1: bisection; 2: grid search)}
}
\value{
the indices of the hypotheses rejected
}
\description{
The main function in SIM paper
}
\examples{
p_1 = rnorm(100)
p_1[1:10]=p_1[1:10]+2
p_1 = 1-pnorm(p_1)
p_2 = rnorm(100)
p_2[1:10]=p_2[1:10]+2
p_2 = 1-pnorm(p_2)
out = SIM(p_1, p_2, alpha = 0.1, SIM_F_0_method=1, method_lambda=1, method_t=2)
}
