% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM_estimated_mu_0_theta}
\alias{SIM_estimated_mu_0_theta}
\title{Function to estimate the mean of SIM p-value under true null}
\usage{
SIM_estimated_mu_0_theta(SIM_p)
}
\arguments{
\item{SIM_p}{single-index SIM_p-value}
}
\value{
\eqn{\hat\mu_0(\theta)}
}
\description{
Function to estimate the mean of SIM p-value under true null
}
