% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM_estimated_sigma_0_theta}
\alias{SIM_estimated_sigma_0_theta}
\title{Function to estimate the sd of the SIM p-value under true null}
\usage{
SIM_estimated_sigma_0_theta(SIM_p)
}
\arguments{
\item{SIM_p}{single-index SIM p-values}
}
\value{
the sd of the SIM p-value
}
\description{
Function to estimate the sd of the SIM p-value under true null
}
