% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM_lambda_selection}
\alias{SIM_lambda_selection}
\title{Function: select lambda used in estimating the null proportion}
\usage{
SIM_lambda_selection(p_theta, SIM_F_0_method, mu_theta, sigma_theta, option)
}
\arguments{
\item{p_theta}{vector of SIM p-values}

\item{SIM_F_0_method}{1: parametric method; 2: nonparametric method}

\item{mu_theta}{\eqn{\mu_0(\theta)}}

\item{sigma_theta}{\eqn{\sigma_0(\theta)}}

\item{option}{options for choosing lambda}
}
\description{
Function: select lambda used in estimating the null proportion
}
