% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM_p_value}
\alias{SIM_p_value}
\title{Function to compute single index p-value}
\usage{
SIM_p_value(p_1, p_2, theta)
}
\arguments{
\item{p_1}{component p-value}

\item{p_2}{component p-value}

\item{theta}{projection direction in \eqn{[0, \pi/2]}}
}
\value{
SIM_p: single index modulated p-value
}
\description{
Function to compute single index p-value
}
