% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SIM_theta_0_selection}
\alias{SIM_theta_0_selection}
\title{Function: optimal selection of theta}
\usage{
SIM_theta_0_selection(p_1, p_2, alpha, SIM_F_0_method, option)
}
\arguments{
\item{p_1}{the primary p-values}

\item{p_2}{the auxiliary p-values}

\item{alpha}{the significance level}

\item{SIM_F_0_method}{the method for estimating the null distribution}

\item{option}{the grid for theta}
}
\value{
the estimated projection direction
}
\description{
Function: optimal selection of theta
}
